Non-GPU environment for publishing docs (#106)
This PR introduces a new `docs` environment to get `pdoc` rendering on
non-GPU runners. Essentially it pulls in CPU-capable versions of
`jaxlib` and friends while rendering. This is needed because `pdoc` uses
dynamic analysis to render docstrings, meaning that it literally imports
and executes the code when it runs. There's one module with `assert`
statements that kills the render process (see
[comment](https://github.com/probcomp/b3d/pull/106#issuecomment-2285031902)),
so that's commented out for now and I captured it as an issue
[here](https://linear.app/chi-fro/issue/GEN-451/remove-assert-statements-in-other-likelihoodspy).