Pixel Kernels (3/2): RGBD Kernels (#152)
While working on the image kernels, I realized that it might be useful
to have a unified class that sample the RGBD values jointly, so I'm
introducing this utility class that takes in an color kernel (from
https://github.com/probcomp/b3d/pull/147) and a depth kernel (from
https://github.com/probcomp/b3d/pull/149) to create an RGBD kernel,
which supports sample and logpdf computation.

This is a general class that's agnostic of the type of RGB and depth
kernels -- it'll pass in the arguments it receives to _both_ of the
kernels during `sample` and `logpdf` call. Since both color and depth
kernels have `*args, **kwargs` as part of their function signature, they
should simply ignore additional arguments that's not relevant to them.

(I also just realized that I was referring to these classes as "kernels"
but the actual class names are `*Distributions`... maybe I should fix
these in a future PR.)

I'm submitting this PR to `main` for now. Once I'm wrapping up my local
changes, I'm going to work on resolving merge conflicts with `gen3d`.

## Test Plan

Similar to the previous PRs, I've added some unit tests to make sure
that the kernels roughly have the behaviors that we expected:

```bash
pytest tests/gen3d/test_pixel_rgbd_kernels.py
```