PixelRGBDDistribution: Additional tests on invalid RGBD values (#159)
This PR includes some additional tests to make sure the behavior of the
`logpdf` method is what we expect when it receives invalid RGBD values
(`[-1., -1., -1., -1.]`). Since the distribution is defined on the pixel
level, it doesn't really expect the observed RGBD to be invalid, so when
we use it to vmap over vertices, we need to handle the case where a
vertex does not hit the image plane manually. I'm also updating the
image kernel slightly to make sure that we can handle those corner
cases.

(note that this is different from the case where a pixel does not have
an associated vertex, which can be handle natively in the pixel
distribution)


# Test Plan

```bash
pytest tests/gen3d/test_pixel_rgbd_kernels.py
```