Docstrings; Add back Log Q scores; Notebook Cleanup (#169)
This adds back in the logq scores in the inference algorithm. It seems
to work pretty well!

Interestingly, on the dataset in the tester.ipynb notebook, having the
logq scores changes the qualitative behavior at frame ~23, when the
camera exposure changes. Without the logq scores, the tracking works
okay through this period, while with the logq scores, the algorithm
loses track of the object.

This warrants further debugging, since it probably indicates the math is
wrong somewehre. I suspect it may have to do with the fact that we are
currently not including scores for pixels that are unobserved. I wonder
if this makes the algorithm too likely to want to resample poses that
have many points off the screen, since the scoring doesn't realize how
much extra randomness this induces. (Put another way, if we include this
score, we should lower the P scores of poses with many points off
screen, since we will have a log(1/(far - near)) term for each pixel
that no point is hitting, and there are more of these for poses where
many points are off screen.)